+++ src/Library.fs
@@ -1,5 +1,0 @@
-﻿namespace RealWorldFable
-
-module Say =
-    let hello name =
-        printfn "Hello %s" name